Hello World DE1 project working.

+++ HelloWorld/Firmware/uart.h
@@ -1,10 +1,10 @@
 #ifndef UART_H
 #define UART_H
 
-/* Hardware registers for a supporting UART to the ZPUFlex project. */
+/* Hardware registers for a supporting UART to the TG68MiniSOC project. */
 
 #define UARTBASE 0xFFFFFFC0
-#define HW_UART(x) *(volatile unsigned int *)(UARTBASE+x)
+#define HW_UART(x) *(volatile unsigned short *)(UARTBASE+x)
 
 #define REG_UART 0x0
 #define REG_UART_RXINT 9
